pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "bitsandbytes>=0.49.2",
    "deepeval>=3.8.4",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "llama-index>=0.14.15",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-retrievers-bm25>=0.6.5",
    "openai-agents>=0.9.1",
    "optimum>=2.1.0",
    "path>=17.1.1",
    "peft>=0.18.1",
    "python-dotenv>=1.2.1",
    "unstructured-inference>=1.2.0",
    "unstructured[all-docs]>=0.21.5",
]
